refactor(sdk): change scss files names, add common files for variables and mixins for sass, use mixins and variables in other scss files
