+++ src/base/_mixins.scss
@@ -1,0 +1,36 @@
+@mixin size($width, $height) {
+  width: $width;
+  height: $height;
+}
+
+@mixin padding($top, $bottom, $left, $right) {
+  padding-top: $top;
+  padding-bottom: $bottom;
+  padding-left: $left;
+  padding-right: $right;
+}
+
+@mixin flexContainer(
+  $direction,
+  $alignment: null,
+  $justify: null,
+  $wrap: null
+) {
+  display: flex;
+
+  @if $direction != null {
+    flex-direction: $direction;
+  }
+
+  @if $alignment != null {
+    align-items: $alignment;
+  }
+
+  @if $justify != null {
+    justify-content: $justify;
+  }
+
+  @if $wrap != null {
+    flex-wrap: $wrap;
+  }
+}
