+++ src/base/_variables.scss
@@ -1,0 +1,3 @@
+$font-type: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+
+$primary-color: #b6a4ff;
