+++ src/controls/GridControl/_grid.scss
@@ -1,3 +1,13 @@
+%modal-button {
+  border-radius: 32px;
+  padding: 16px;
+  padding-left: 48px;
+  padding-right: 48px;
+  border-color: $primary-color;
+  font-size: 16px;
+  border-style: double;
+}
+
 .mapbox-control-cell {
   &-popup {
     position: absolute;
@@ -18,8 +28,7 @@
 
 .mapbox-control-grid-selector-modal {
   &-root {
-    width: 100%;
-    height: 100%;
+    @include size(100%, 100%);
     display: none;
     position: fixed;
     z-index: 1;
@@ -27,7 +36,6 @@
     top: 0;
     overflow: auto;
     padding-top: 100px;
-    background-color: black;
     background-color: rgba(0, 0, 0, 0.4);
   }
   &-content {
@@ -42,17 +50,14 @@
     border-radius: 16px;
   }
   &-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
+    @include flexContainer(null, center, space-between);
     padding: 6px 30px;
     background-color: #fff;
     color: #000;
     border-radius: 16px;
   }
   &-body {
-    display: flex;
-    align-items: center;
+    @include flexContainer(null, center);
     padding: 30px 30px;
   }
   &-divider {
@@ -82,45 +87,26 @@
     margin-bottom: 36px;
   }
   &-buttons-container {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
+    @include flexContainer(null, center, space-between);
     padding: 30px 30px;
   }
-  &-select{
-    padding:8px;
+  &-select {
+    padding: 8px;
     padding-right: 16px;
   }
-  &-close-button{
+  &-close-button {
     background: transparent;
     border: transparent;
     padding-top: 8px;
   }
-  &-select-button{
-    border-radius: 32px;
-    padding:16px;
-    padding-left: 48px;
-    padding-right: 48px;
-    background-color: white;
-    background-color:#B6A4FF;
-    border-color: #B6A4FF;;
+  &-select-button {
+    @extend %modal-button;
+    background-color: $primary-color;
     color: white;
-    font-size: 16px;
-    border-style: double;
   }
-  &-cancel-button{
-    border-radius: 32px;
-    padding:16px;
-    padding-left: 48px;
-    padding-right: 48px;
+  &-cancel-button {
+    @extend %modal-button;
     background-color: white;
-    color:#B6A4FF;
-    border-color: #B6A4FF;
-    font-size: 16px;
-    border-style: double;
+    color: $primary-color;
   }
-}
-
-.mapbox-cancel-button {
-  background-color: red;
 }
