--- src/controls/GridControl/components/ControlButton.scss
@@ -1,15 +1,0 @@
-.mapbox-control {
-  button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &.-active {
-      background: rgba(0, 0, 0, 0.03);
-
-      svg {
-        fill: #4264fb;
-      }
-    }
-  }
-}
