+++ src/controls/components/_controlButton.scss
@@ -1,0 +1,9 @@
+.mapbox-control {
+  button {
+    @include flexContainer(null, center, center);
+
+    &.-active {
+      background: rgba(0, 0, 0, 0.03);
+    }
+  }
+}
