--- src/index.scss
@@ -1,3 +1,0 @@
-@import "controls/GridControl/components/ControlButton.scss";
-@import "controls/GridControl/GridControl";
-@import "controls/MapTilesControl/MapTilesControl.scss";
